refactor: Customer category
