+++ app/src/main/java/com/mifos/apache/fineract/injection/component/ActivityComponent.java
@@ -2,11 +2,11 @@ package com.mifos.apache.fineract.injection.component;
 
 import com.mifos.apache.fineract.injection.PerActivity;
 import com.mifos.apache.fineract.injection.module.ActivityModule;
-import com.mifos.apache.fineract.ui.online.createcustomer.customeractivity.CreateCustomerActivity;
-import com.mifos.apache.fineract.ui.online.createcustomer.formcustomeraddress.FormCustomerAddressFragment;
-import com.mifos.apache.fineract.ui.online.customer.CustomersFragment;
+import com.mifos.apache.fineract.ui.online.customer.createcustomer.customeractivity.CreateCustomerActivity;
+import com.mifos.apache.fineract.ui.online.customer.createcustomer.formcustomeraddress.FormCustomerAddressFragment;
+import com.mifos.apache.fineract.ui.online.customer.customerlist.CustomersFragment;
 import com.mifos.apache.fineract.ui.online.customerdeposit.CustomerDepositFragment;
-import com.mifos.apache.fineract.ui.online.customerdetails.CustomerDetailsFragment;
+import com.mifos.apache.fineract.ui.online.customer.customerdetails.CustomerDetailsFragment;
 import com.mifos.apache.fineract.ui.online.customerloans.CustomerLoansFragment;
 import com.mifos.apache.fineract.ui.online.debtincomereport.DebtIncomeReportFragment;
 import com.mifos.apache.fineract.ui.online.depositdetails.CustomerDepositDetailsFragment;
