+++ app/src/main/java/com/mifos/apache/fineract/ui/online/customer/createcustomer/customeractivity/CreateCustomerContract.java
@@ -1,4 +1,4 @@
-package com.mifos.apache.fineract.ui.online.createcustomer.customeractivity;
+package com.mifos.apache.fineract.ui.online.customer.createcustomer.customeractivity;
 
 import com.mifos.apache.fineract.data.models.customer.Customer;
 import com.mifos.apache.fineract.ui.base.MvpView;
