+++ app/src/main/java/com/mifos/apache/fineract/ui/online/customer/createcustomer/formcustomeraddress/FormCustomerAddressFragment.java
@@ -1,4 +1,4 @@
-package com.mifos.apache.fineract.ui.online.createcustomer.formcustomeraddress;
+package com.mifos.apache.fineract.ui.online.customer.createcustomer.formcustomeraddress;
 
 import android.app.Activity;
 import android.content.Context;
@@ -21,7 +21,7 @@ import com.mifos.apache.fineract.data.models.customer.Address;
 import com.mifos.apache.fineract.ui.base.MifosBaseActivity;
 import com.mifos.apache.fineract.ui.base.MifosBaseFragment;
 import com.mifos.apache.fineract.ui.base.Toaster;
-import com.mifos.apache.fineract.ui.online.createcustomer.OnNavigationBarListener;
+import com.mifos.apache.fineract.ui.online.customer.createcustomer.OnNavigationBarListener;
 import com.stepstone.stepper.Step;
 import com.stepstone.stepper.VerificationError;
